refactor(advanced): 페이지 컴포넌트 분리
- 관리자 및 카트 페이지를 pages/로 이동
