+++ src/advanced/pages/admin/index.ts
@@ -1,0 +1,1 @@
+export { AdminPage } from './ui/AdminPage';
