+++ src/advanced/pages/cart/index.ts
@@ -1,0 +1,1 @@
+export { CartPage } from './ui/CartPage';
